% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condaenv.R
\name{condaenv}
\alias{condaenv}
\title{Global conda environment}
\usage{
condaenv(conda = NULL)
}
\arguments{
\item{conda}{Conda environment to be use. If \code{NULL}, will return current
global conda environment, if \code{NA}, will remove the global conda environment
settting. If a string, will set the global conda environment.}
}
\description{
All python related function will accept \code{conda} argument, if it's \code{NULL},
then the global conda environment will be used, if it's \code{NA}, no conda
environment will be used, otherwise it must be a string indicates the conda
environment name.
}
