% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysc-package.R
\docType{package}
\name{pysc-package}
\alias{pysc}
\alias{pysc-package}
\title{pysc: Miscellaneous Python functions for Single Cell analysis}
\description{
Miscellaneous Python functions for Single Cell analysis.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
