% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccoda.R
\name{sccoda}
\alias{sccoda}
\title{Bayesian model for compositional single-cell data analysis}
\usage{
sccoda(data, formula, reference = "automatic", ...)
}
\arguments{
\item{data}{A Python \code{AnnData} object returned by \link{sccoda_data}.}

\item{formula}{Patsy-style formula for building the covariate matrix.
Categorical covariates are handled automatically, with the covariate value of
the first sample being used as the reference category. To set a different
level as the base category for a categorical covariate, use
\verb{C(<CovariateName>, Treatment('<ReferenceLevelName>'))}.}

\item{reference}{The \code{scCODA} model requires a cell type to be set as the
reference category. However, choosing this cell type is often difficult. A
good first choice is a referenece cell type that closely preserves the
changes in relative abundance during the compositional analysis.}

\item{...}{
  Arguments passed on to \code{\link[=sccoda_sampling]{sccoda_sampling}}
  \describe{
    \item{\code{sampling}}{There are three different MCMC sampling methods available for
\code{scCODA}. Default: "hmc".
\itemize{
\item \code{hmc}: Hamiltonian Monte Carlo (HMC) sampling,
\item \code{da}: HMC sampling with Dual-averaging step size adaptation.
\item \code{nuts}: No-U-Turn sampling.
}}
    \item{\code{num_results}}{MCMC chain length. Default: 20000 for "hmc" and "da"
sampling; 10000 for "nuts" sampling.}
    \item{\code{num_burnin}}{Number of burnin iterations. Default: 5000.}
    \item{\code{num_adapt_steps}}{Length of step size adaptation procedure.}
    \item{\code{num_leapfrog_steps}}{HMC leapfrog steps (default 10).}
    \item{\code{max_tree_depth}}{Maximum tree depth (default 10).}
    \item{\code{step_size}}{Initial step size (default 0.01).}
    \item{\code{verbose}}{If true, a progress bar is printed during MCMC sampling}
  }}
}
\description{
scCODA is a toolbox for statistical models to analyze changes in
compositional data, especially from single-cell RNA-seq experiments. Its main
purpose is to provide a platform and implementation for the scCODA model,
which is described by \href{https://www.nature.com/articles/s41467-021-27150-6}{Büttner, Ostner et al.}.
}
