% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccoda.R
\name{sccoda_analysis}
\alias{sccoda_analysis}
\title{Performing compositional analysis with scCODA.}
\usage{
sccoda_analysis(data, formula, reference)
}
\arguments{
\item{data}{A Python \code{AnnData} object returned by \link{sccoda_data}.}

\item{formula}{Patsy-style formula for building the covariate matrix.
Categorical covariates are handled automatically, with the covariate value of
the first sample being used as the reference category. To set a different
level as the base category for a categorical covariate, use
\verb{C(<CovariateName>, Treatment(‘<ReferenceLevelName>’))}.}

\item{reference}{The \code{scCODA} model requires a cell type to be set as the
reference category. However, choosing this cell type is often difficult. A
good first choice is a referenece cell type that closely preserves the
changes in relative abundance during the compositional analysis.}
}
\description{
Performing compositional analysis with scCODA.
}
