% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccoda.R
\name{sccoda_data}
\alias{sccoda_data}
\alias{sccoda_data.data.frame}
\alias{sccoda_data.SummarizedExperiment}
\title{Prepare a Anndata for scCODA}
\usage{
sccoda_data(data, ...)

\method{sccoda_data}{data.frame}(
  data,
  covariates,
  celltype,
  sample = NULL,
  counts = NULL,
  ...,
  conda = NULL
)

\method{sccoda_data}{SummarizedExperiment}(
  data,
  covariates,
  celltype = "label",
  sample = "Sample",
  counts = NULL,
  ...,
  conda = NULL
)
}
\arguments{
\item{data}{A data.frame. Alternatively, a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment} or
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}, in this
way, the \code{colData} will be used to build \code{AnnData}.}

\item{...}{Additional arguments passed to specific methods.}

\item{covariates}{Other columns to be added into final object.}

\item{celltype}{A string define the cell types for each cell (in rows) if
\code{sample} is not \code{NULL}, or an atomic character define the columns of counts
for each cell if \code{sample} is \code{NULL}.}

\item{sample}{A string. If specified, \code{data} is expected to be in long
format, with \code{celltype} defining the column information of cell types. If
\code{counts} is \code{NULL}, the internal function will count each cell type.
Otherwise, the column in \code{counts} will be used directly and transformed into
a wide data.frame. If \code{NULL}, \code{data} is expected to be in wide format, with
columns specified in \code{celltype} containing the count data for each cell type.}

\item{counts}{A string define the column of sample information, only used
when \code{sample} is not \code{NULL}.}

\item{conda}{A character vector of preferred python environment names to
search for and use. Details see \link{condaenv}.}
}
\value{
A Python \code{AnnData} object
}
\description{
Prepare a Anndata for scCODA
}
