% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccoda.R
\name{sccoda_datasets}
\alias{sccoda_datasets}
\title{Import datasets from scCODA}
\usage{
sccoda_datasets(name, conda = NULL)
}
\arguments{
\item{name}{The name of dataset to extract.}

\item{conda}{A character vector of preferred python environment names to
search for and use. Details see \link{condaenv}.}
}
\description{
Import datasets from scCODA
}
