% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccoda.R
\name{sccoda_sampling}
\alias{sccoda_sampling}
\title{Performing MCMC sampling}
\usage{
sccoda_sampling(
  model,
  num_results = NULL,
  num_burnin = 5000L,
  num_adapt_steps = NULL,
  num_leapfrog_steps = 10L,
  max_tree_depth = 10L,
  step_size = 0.01,
  verbose = TRUE,
  sampling = NULL
)
}
\arguments{
\item{model}{Result returned by \link{sccoda_analysis}}

\item{num_results}{MCMC chain length. Default: 20000 for "hmc" and "da"
sampling; 10000 for "nuts" sampling.}

\item{num_burnin}{Number of burnin iterations. Default: 5000.}

\item{num_adapt_steps}{Length of step size adaptation procedure.}

\item{num_leapfrog_steps}{HMC leapfrog steps (default 10).}

\item{max_tree_depth}{Maximum tree depth (default 10).}

\item{step_size}{Initial step size (default 0.01).}

\item{verbose}{If true, a progress bar is printed during MCMC sampling}

\item{sampling}{There are three different MCMC sampling methods available for
\code{scCODA}. Default: "hmc".
\itemize{
\item \code{hmc}: Hamiltonian Monte Carlo (HMC) sampling,
\item \code{da}: HMC sampling with Dual-averaging step size adaptation.
\item \code{nuts}: No-U-Turn sampling.
}}
}
\description{
Performing MCMC sampling
}
