% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{pysc_tidy}
\alias{pysc_tidy}
\title{Turn an object into a data.table}
\usage{
pysc_tidy(x, ...)
}
\arguments{
\item{x}{A python object.}

\item{...}{Additional arguments to tidying method.}
}
\description{
Turn an object into a data.table
}
