% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccoda.R
\name{sccoda_datasets}
\alias{sccoda_datasets}
\title{Import datasets from scCODA}
\usage{
sccoda_datasets(name)
}
\arguments{
\item{name}{The name of dataset to extract.}
}
\description{
Import datasets from scCODA
}
